Back/requirements.txt
Django==4.1.7
openai
djangorestframework
django-cors-headers
psycopg2
python-dotenv
django-database-url
gunicorn
djangorestframework-simplejwt
drf-yasg
setuptools